Add files via upload
Q1 - Given an array of numbers, use the map function with an arrow function to square each element of the array.

Q2 - Write a JavaScript function that takes a student's score as a parameter and returns their grade (A, B, C, D, or F) using a series of ternary operators..

Q3 - Create an object representing a car with properties like Company name, model, and year. Write a function to change the car's year property. Also use object destructuring to extract and print the car's Model and Year.

Q4 - Given an array of numbers, use the filter function to create a new array containing only the prime numbers.

Q5 -  State different use cases of map, filter and reduce functions.

Q6 - Write an asynchronous function using async-await  to fetch data from an API (you can use the JSONPlaceholder API) and log the result.

Q7 - Define a nested object representing a person with properties like name, address, and contact. Use optional chaining to safely access the person's phone number, even if the contact property is missing.